P2P: Do not change SSID during GO negotiation.
If GO Negotiation Request (or in theory, also GO Negotiation Response)
frame is delivered multiple time for processing, the SSID of the group
could end up getting changed. This could result in possible issues if
the peer ended up using different SSID. To avoid this, make sure the
SSID does not get changed unless the negotiation is for a new group.

Change-Id: I59f0c26f7af3d6b120e63c658c80de999cdd2f29