nl80211: Consider P2P when changing vif type and do not re-set iftype.
Added support for P2P interfaces when adding a new interface. However,
it didn't handle the case in which the same interface is being used and
its type is being changed. Add support for this case.

Consequently, when doing "ap_scan_as_station" we now need to save the
actual AP interface type (AP/P2P GO) in order to restore it properly.
For that, change ap_scan_as_station type from int to nl80211_iftype, and
set it to NL80211_IFTYPE_UNSPECIFED when not used.There is no need to
force the interface into station mode when wpa_supplicant adds a new
interface (e.g., P2P group) with the correct iftype.

Change-Id: Iefa3cc960e9394afe218b5eb03e00be8c0d47e2e