Wifi station scan and connect fixes.
nl80211: Work around EALREADY from connect command:
cfg80211 does not currently allow NL80211_CMD_CONNECT when there
is already an association. This can result in wpa_supplicant being
unable to request a connection if the interface is already in
connected state. Work around this by requesting disconnection if
the connect command fails with EALREADY error. This allows
wpa_supplicant to recover from the unexpected state and complete
the connect request on next attempt.

Change-Id: I2ac0ae52f37d9fe48122247afe2933794cf184b2