nl80211: Filter unexpected interface added/up events.
It looks like a RTM_NEWLINK event claiming the interface to be UP is
delivered just before removing an interface after having first indicated
that the interface was going down/removed. Ignore this event if the
interface is not present anymore at the moment the event is processed.
This fixes issues where an interface that was re-added after being
removed did not get reconfigured properly.

Change-Id: I8579e4cba23bf18725e6f5de97fc796821e8bde5
Signed-hostap: Jouni Malinen <j@w1.fi>
intended-for: hostap-1