wpa_supplicant: changes to support WiFi Display
Made changes so that wpa_supplicant and wpacli programs support
WiFi Display operation as specified in version 1.29 of the WiFi Display
Specification published by WiFi display alliance. This Adds support
of an API that allows wpa_supplicant clients to set and retrieve WFD
parameters of the peer devices on the network. This also adds support
for parsing and construction of WFD IEs in wpa_supplicant.

Change-Id: I47750bddeded0f0c4553fdda31ed42696a4ec080