Add os_realloc_array() wrapper
This function can be used as a wrapper for os_realloc(ptr, nmemb * size)
when a reallocation is used for an array. The main benefit over
os_realloc() is in having an extra check to catch integer overflows in
multiplication. This is similar to the os_zalloc() to os_calloc() move.

(cherry picked from commit 733c72e91a052613ed307eca9cf5c42983f5d569)

Signed-hostap: Jouni Malinen <j@w1.fi>
Change-Id: Ifeaa6aca9ae6ca355cc8db5c906cbd0fdfe8062b