Interworking: Do not reconnect without "auto" parameter
When there was no credential match, but an enabled network block matched
with a scan result, wpa_supplicant reconnected at the end of
interworking_select command even if "auto" parameter was not used. Fix
this by running the reconnect only if requested to automatically select
a network.

Change-Id: Ia13a46b87016c908e4b0ed70df9629674c474223
Signed-hostap: Jouni Malinen <j@w1.fi>