Interworking: Fix home SP check with real SIM card
The NAI building routine assumed that the credential included the IMSI,
but that is not the case when using a real SIM card. Build the NAI based
on the IMSI read for the card in such a case.

Change-Id: Id8bc145fbd12f390b7706a9bd5f2872ac82b039d
Signed-hostap: Jouni Malinen <j@w1.fi>
(cherry picked from commit 8fb33b0fccbadab53550ab91975fe6ba8295a69e)