Print control interface commands in easier format
The ASCII hexdump is somewhat difficult to search for (especially on
Android builds), so make the debug log easier to parse by printing the
full control interface command as a text string. In addition, use
wpa_dbg() to get the interface name printed so that multi-interface
cases can be debugged.

Change-Id: I7744787e2fc2ff021b224bbd7ef415d1ba7922ad
Signed-hostap: Jouni Malinen <j@w1.fi>