wpa_supplicant_8:Allow all channels with multi-channel concurrency
If the driver indicates support for multi-channel concurrency, change
the p2p_connect behavior to not force the current operating channel, but
instead, just mark it as preferred for GO Negotiation. This change
applies only for the case when the freq parameter is not used with the
p2p_connect command.

(cherry picked from commit 3c6ed923748fea13bad1470abb67b49f5500057b)

Change-Id: I3c85db312278875052ca7c258d5c07fe42f408be