wpa_supplicant_8: Do not inform other virtual interfaces of scan results in all cases
If a connection operation is started on an interface based on scan
results, other virtual interfaces should not be information about the
results to avoid potential concurrent operations during the association
steps. Since the sibling notification of scan results received was added
as an optimization, skipping it for this type of cases is the simplest
way of avoiding unnecessary concurrent operations.

Conflicts:

	wpa_supplicant/events.c

(cherry picked from commit a16963f300a4763cc77e5f6f2da409bfae027042)

Change-Id: I2a47eb880ae24b00b35edffe48d58978250c4177