wpa_supplicant_8: Enhancing "Optimize post-WPS scan based on channel used during provisioning"
Commit 17a4734dc4d28ef070dfa227b422ed84794f48c5 shall optimize the
post-WPS scan based on the channel used during the provisioning only
when the EAP-Failure frame is received . In cases where the EAP-Failure
frame is missed , this optimization shall not happen resulting in the
scan in all the channels thus consuming more time for the connection.
This commit enhances this by storing the requisite information after
the M8 frame is received.

CRs-fixed: 429147
Change-Id: I86c0ae2278db00fe35827ca2c70af1cec82032b5
(cherry picked from commit c7f3a7ae14599868b72809700378ca83b12b9410)