wpa_supplicant_8: WPS:Update configuration file after re-enabling networks
Network blocks are disabled during a WPS provisioning step and 10
seconds after this to allow the newly provisioned network to be
selected. The disabled=1 flag gets written to the configuration file
when the credential is added since it happens during this process.
Update the file again after the networks have been re-enabled to avoid
leaving the configuration file into state that is not consistent with
the running configuration.

Change-Id: I877c84efb060a9ffa156e0bb4bd1abcdc78d523b
CRs-fixed: 398860
(cherry picked from commit 4c12d57ad2853d0c8802ecac8fff3282735ad33c)