P2P: Share a single function for GO channel selection.
The exact same mechanism was used for determining the operating channel
at the device that becomes the GO regardless of whether this was
triggered by reception of GO Negotiation Request of Response frame. Use
a shared function to avoid duplicated implementation and potential
differences in the future.

Change-Id: Idf367dc3fbbf5e0f828f53f272ee2f24c14ecd7b
Signed-hostap: Jouni Malinen <j@w1.fi>