P2P: Allow all channels with multi-channel concurrency.
If the driver indicates support for multi-channel concurrency, change
the p2p_connect behavior to not force the current operating channel, but
instead, just mark it as preferred for GO Negotiation. This change
applies only for the case when the freq parameter is not used with the
p2p_connect command.

Change-Id: I9db61b0bc03eb1aeedec8a4d76274791f050e51f
Signed-hostap: Jouni Malinen <jouni@qca.qualcomm.com>