P2P: Share a single function for GO channel selection.
The exact same mechanism was used for determining the operating channel
at the device that becomes the GO regardless of whether this was
triggered by reception of GO Negotiation Request of Response frame. Use
a shared function to avoid duplicated implementation and potential
differences in the future.

Change-Id: Ibbe89db3e967b4fde742ac1eedbdfc415fbee041
Signed-hostap: Jouni Malinen <j@w1.fi>