P2P: Always re-select operating channel if not hard coded.
Since the operating channel is randomly set to 1/6/11 on init, which is
commonly included in the channel intersection, we were effectively
ignoring the set of P2P preferred channels when trying to improve
channel selection after having received peer information. Fix this by
trying to get the best channel we can, unless the user hard coded the
operating channel in the configuration file or p2p_connect command. Fall
back to the initial randomly selected channel if a better one cannot be
chosen.

Change-Id: Ic92774d96d6ae19efed00d9e96a5941cdad0c25a
Signed-hostap: Arik Nemtsov <arik@wizery.com>