P2P: Use peer's channel list to limit GO freq on invitation
Peer device includes its list of allowed operating channels in the
Invitation Response frame. When we are becoming the GO, use that list
from the peer to filter out acceptable channels to avoid selecting a
channel that the peer is unable to use.

Change-Id: I4e69bc0fea4cdba7dfb7f03a17e7ee8e28ea5897
Signed-hostap: Jouni Malinen <jouni@qca.qualcomm.com>
CRs-Fixed: 453053