Do not add BSS to blacklist on local disconnection request
There is no point in marking a BSS temporarily blacklisted based on a
connection failure or disconnection case if that happens as a result of
a local request to disconnect. The blacklist entry could result on
unexpected BSS getting selected on the next connection attempt. In
addition, the code to try to find another BSS within the ESS could
result in scanning a single channel on the next attempt. Fix these
issues by handling the connection failure events only if we are not in
disconnected state (i.e., would try to reconnect after this
automatically).

Signed-hostap: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: 0cdb93fe9ffe2bfb125f6f758f837a09e7762155
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: If12c83ca1c581a901a43465f91a31687aa95174a
CRs-fixed: 487798