nl80211: Add "interface unavailable" event API
Interface unavailable API command is used by driver to indicate that
the interface is not available to userspace. The driver may have
encountered errors that has forced the driver to report this event.
When the driver is in P2P-Go mode, and when it goes thru automatic
recovery from firmware crashes, it uses this interface to notify the
userspace that the group has been deleted.

Change-Id: I830d3a74abb440a71a610d62438f49fc64bc7518