Configure beacon interval for IBSS command
wpa_supplicant already allowed beacon interval to be configured for AP
mode operations, but this was not passed to the driver for IBSS even
though the same parameter can used for that case. Add this for the
nl80211 driver interface to allow beacon interval to be controlled for
IBSS as well.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: 8f05577d1137fe8753efb5278efd3d6a942e5b20
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: I25e7b0786a39cdd31c7da29cd59f2fc3bf0adf0d
CRs-fixed: 616006