P2P: Allow GO Negotiation Request to update peer capability attribute
Previously, GO Negotiation Request frame was used to update a peer entry
if,(i) a Probe Request from that peer had been received, or (ii) the peer
is discovered based on receiving Provision Discovery Request frame from
it and no Probe Request frame. However, it is possible that the peer is
discovered based on a Probe Rsp. received from it, then the capability
attribute of the peer changes, e.g., peer ends GO P2P operation. In this
case, if a GO Negotiation frame is received from the peer, the capability
attribute in the GO-NEG-REQ frame will be different from the one in peer
entry within P2P context. This will cause device to use an incorrect
"join" flag when connecting to the peer device

Fix this by allowing the GO Negotiation Request frame to update
the device and group capabilities of the peer entry if the current peer
entry has a capability attribute value different from the one in the
received GO Negotiation Request.

Change-Id: If2ccfc9caca76457b3d049d86279ae7b348e2eb9
CRs-fixed: 622170