Set the station to authorized on assoc event for open authentication
This was done in handle_assoc_cb() in ieee802_11.c for drivers that use
hostapd SME/MLME. However, drivers that include SME/MLME implementation
do not use that function and the STA flag needs to be set when
processing the association notification. This is needed to get the STA
entry into showing the proper authorized state and to get the
AP-STA-CONNECTED/DISCONNECTED events on the control interface.

CRs-fixed: 608957
Git-commit: 113318ad574a47d41dcb1f160d11500e2d140d7d
Git-repo : git://w1.fi/srv/git/hostap.git
Change-Id: I7effbe5bfdb4838b4cc8c79a745dc5a0ad4fe78f
Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>