Add SIM identifier to the network profile and cred block
This allows the specific SIM to be identified for authentication
purposes in multi-SIM devices. This SIM number represents the index of
the SIM slot. This SIM number shall be used for the authentication using
the respective SIM for the Wi-Fi connection to the corresponding
network.

CRs-Fixed: 631880
Git-commit: 13f6a07efc3e737d8fcb8f7014cb961ce77e64ee
Git-repo : git://w1.fi/srv/git/hostap.git
Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>

Change-Id: I82becee92043e4da5a672353b1fdac6cf0bc531a