eap_proxy: Check for modem support at the run time
Some of the targets do not support modem. But EAP-SIM feature
requires modem support. The modem initialization will fail on
these targets leading to supplicant initialization failure.
To mitigate this, check for modem presence at the run time and
avoid the proxy initialization if modem support is not present.

Change-Id: I6cf6a5b15886f6df09bae0026b988ed9b7d26e3f
CRs-Fixed: 689647