hostapd: Add vendor specific VHT extension for the 2.4 GHz band
This allows vendor specific information element to be used to advertise
support for VHT on 2.4 GHz band. In practice, this is used to enable use
of 256 QAM rates (VHT-MCS 8 and 9) on 2.4 GHz band.

This functionality is disabled by default, but can be enabled with
vendor_vht=1 parameter in hostapd.conf if the driver advertises support
for VHT on either 2.4 or 5 GHz bands.

Signed-off-by: Yanbo Li <yanbol@qti.qualcomm.com>
Git-commit: e7d0e97bdbdc996564f06b382af3d5a5164a8fb3
Git-repo : git://w1.fi/srv/git/hostap.git
Change-Id: I172cdbda6b547d38e9842f3974eee8c763210668
CRs-Fixed: 1043973