Fix SCAN control interface command error cases
Update the scan parameters in wpa_s only in case the scan command is
going to be executed. In other words, do not change the parameters for
an ongoing scan (the SCAN command is rejected with FAIL-BUSY) or if any
of the parameters is invalid.

CRs-Fixed: 867249
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 43a66ecba3f1ad41ab36f82d9e979ca5cef919d5
Signed-off-by: Jouni Malinen <j@w1.fi>

Change-Id: I43f41078e11afd14cd20031a7eb9a4038a548f0e