Ignore too long SSID element value in parser
The SSID element is defined to have a valid length range of 0-32. While
this length was supposed to validated by the users of the element
parser, there are not really any valid cases where the maximum length of
32 octet SSID would be exceeded and as such, the parser itself can
enforce the limit as an additional protection.

CRs-Fixed: 867249
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 05e46a944ac6f5667b180e8ff49793e3c45ae6dd
Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>

Change-Id: I10fd5705d24ea2b55a5671d4f693cc263030cac0