Don't write to wpa_supplicant.conf directly
There is a chance that wpa_supplicant may get killed during
the time it is writing config file. If this happens, user
information like SSIDs and passwords can be lost forever.

This change works around that by writing config to a
temporary file and then renaming the file to the correct name.

CRs-Fixed: 830227
Change-Id: I55f56f2bca7ecc41af954d43356ea7c8e3637310
Signed-off-by: Vinit Deshpande <vinitd@google.com>