Add libwpa_client build option to use a dynamic library
Add support to compile libwpa_client.so. This can be used by external
programs to interact with the wpa_supplicant control interface. Also
rename the static version of this library to libwpa_client.a to be
consistent with the name used previous in Android builds.

CRs-Fixed: 838597
Git-commit: 701d9729b3a7ff574c9ff29048b82c00df1a6a68
Git-repo : git://w1.fi/srv/git/hostap.git
Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>

Change-Id: I47a837be40cb7751e64d4d6b11cf51f253764463