Advertize cross connection p2p ie's to WLAN AP on a non p2p interface
commit 1c2aa04c96626f5b000d167bb5274a8c39b7dac2 addresses to not add
the p2p ie's on a non p2p interface.AP's functioning as a p2p
manager shall need the cross connection capability of the
station. This commit ensures the device sends this capability
for a WLAN AP on a query to get the same.
This is also required to address the certification requirements.

Change-Id: I62f3fa432451d15296e0fe82c439a286c565c226
CRs-fixed: 836613