Add support to request a scan with specific SSIDs
Support a request to scan specific SSIDs given by user with the SCAN
command. The SSID list can be suffixed to the scan command as follows.
For example, if SSIDs "ABC" and "abc123" need to be specifically
scanned, the command should be "SCAN ssid 414243 ssid 616263313233". The
value of the SSID is passed in hexadecimal representation.

CRs-Fixed: 867249
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: a80651d06774dfd93b51d1f8b71345956724807a
Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>

Change-Id: I04c73e4d1c13d23ff332ebf5d2733f275d444faa