EAP-pwd server: Fix Total-Length parsing for fragment reassembly.
The remaining number of bytes in the message could be smaller than the
Total-Length field size, so the length needs to be explicitly checked
prior to reading the field and decrementing the len variable. This could
have resulted in the remaining length becoming negative and interpreted
as a huge positive integer.
In addition, check that there is no already started fragment in progress
before allocating a new buffer for reassembling fragments. This avoid a
potential memory leak when processing invalid message.

CRs-Fixed: 833592
Git-commit: 3035cc2894e08319b905bd6561e8bddc8c2db9fa
Git-repo : git://w1.fi/srv/git/hostap.git
Signed-off-by: Jouni Malinen <j@w1.fi>

Change-Id: I458174b88195f09a712c775896e6efa8f2791bc0