RSN: Remove check for proactive_key_caching while setting PMK offload
wpa_sm_key_mgmt_set_pmk() was checking for proactive_key_caching to be
enabled before setting the PMK to the driver. This check is not required
and would mandate configuration setting of okc or proactive_key_caching
for cases which were not necessary.

Signed-off-by: Amarnath Hullur Subramanyam <amarnath@qca.qualcomm.com>
Git-commit: 89a11ad38f9faace28213653eaecbad98460c9e1
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: I599c8d8e0f5595e4a26ecbe6486fcb571e9043c1
CRs-Fixed: 940433