P2P: Add a separate pointer to the P2P Device instance
In many places in the code there was a reference to wpa_s->parent to get
from group interface to p2p_dev interface. These places can break if
P2P_DEVICE interface would need to be used with the primary interface as
the group interface, since the parent of the primary interface points to
itself and not the p2p_dev interface.

Fix this by adding a separate "p2pdev" pointer to wpa_supplicant,
it will be the same as parent pointer in most cases but whenever
the primary interface is used as a group interface, change it to
point to the correct p2p_dev interface.

Change-Id: Ice5a97a51e245b0866f297bad976c1016bf1e4ab
Signed-off-by: Lior David <qca_liord@qca.qualcomm.com>
Git-commit: ba307f8528b1242c931a57feb21cd30286126916
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 985990