P2P: Add optional op_class argument to P2P_SET listen_channel
The existing implementation in p2p_ctrl_set used a hard-coded operating
class 81 which is only suitable for the social channels in the 2.4 GHz
band, and will not work for the social channel in the 60 GHz band.
Extend this by adding an optional op_class argument to P2P_SET
listen_channel. If not specified, use the default value of 81 to match
existing behavior.

Change-Id: I6a60a47c8dea72059fd32bca94d4e28e06a5936e
Signed-off-by: Lior David <qca_liord@qca.qualcomm.com>
Git-commit: dfe0745c9a8246400316879be0115cc119c4882f
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 985990