scan: Extend AP scan timeout workaround for emulation platform only
For FPGA platform, firmware is incurring an additional delay in completing
scans. Therefore, we need to extend the AP scan timeout in the supplicant so
it doesn't fire a new scan request before scan results of the previous scan
are made available by firmware. The changes are enclosed by
QCA_WIFI_3_0_EMU_SUPPLICANT flag only when EOS or Adrastea config is defined
in the defconfig.

Change-Id: Ic500507616bad3caf3b39ffa39f6839a81bec3ac
CRs-Fixed: 985210